Code to create line divider between content and sidebar and for this line to extend top to bottom regardless of the height of either content or sidebar
This solution is using outlined in
http://blog.credera.com/topic/technology-solutions/full-height-column-layout-with-bootstrap/

However if this proves to be an issue in regards to the footer section of the site then I could revisit the idea using border on both the sidebar and content area and then use negative margins to overlap them.

This way one margin will always extend full available height.

Signed-off-by: Ryan Holder <ryan@ryanholder.com>