Add in our own structural wrap called .col-wrap
```
<div class="wrap">
	<div class="col-wrap">…</div>
</div>
```

We are using the default Genesis structural wrap (.wrap) as our layouts (.container) elements. Containers control the max and min widths of the layout based on device being used.

However we needed an extra structural wrap that acts as the "row" identifier which is crucial to the grid layout we are making use of.

Therefore I am adding in our .col-wrap div as shown above.

As of now the only minor gripe is that I have to add our div by inserting another div with class .wrap and then rename the default .wrap to .col-wrap.

Reason for this is explained more in the comments however in a nutshell it is because we can only execute adding html before the default wrap gets added.

Signed-off-by: Ryan Holder <ryan@ryanholder.com>