Rename Image Dimension to Image Size (px)
No need to use a long word like Dimension when Size will do. It's a more
common word, and probably easier to translate into other languages.
Includes a unit within the label, to indicate what the unit is, and that
it doesn't need to be included within the field value itself.

Since we're pre-release, we can afford to rename the field key to match.

Fixes #195