Rename Button references to Link
The term "button" should not be used in the Featured Content widget
form, since the output is not a (and nor should it be), so even if it
currently happens to look like a button in FortyTwo, it might not look
like it in a forked or customised child theme.

Since we're pre-release, we can afford to rename the fields to match.

Fixes #195