Use customised code standards ruleset
This is to avoid the false positives that are given from the WordPress
Coding Standards sniffs regarding escaping.

Some of the results from this sniff are correct, and have been fixed,
but the remaining ones are regarding escaping content in HTML context
(not HTML attribute context), when we may not want to escape_html().