Re-organize all the things
Root directory now split into assets, src and tests.

* assets - subdirectories of bower (populated by grunt task), composer
(will be populated via grunt task, but current contains all of wpthumb
project), and forsitethemes (custom styles, scripts and images). This
directory is not included in the final .zip and effectively lives
outside of the theme root.
* src - the theme root, containing typical theme PHP, style.css etc. Has
subdirectories of images/, js/, fonts/, lib/ and templates/. The first
three are populated with dependencies from the assets directory via
grunt tasks. The copying also images inside lib/admin/ and wpthumb to
lib/wpthumb. If running this theme locally, then the symlink should now
be to this directory, not the project root.
* tests - no change here.

Use of copy task means that whatever can't be filtered from bower or
composer package retrievals i.e. only the five .php files for wpthumb
and not the FontAwesome.otf for font-awesome fonts (applying a filter
in bower.json seems to affect all directories of that type, something
broken in grunt-bower-task).

Since the a smaller superset of actual files used are now in assets
directory, the large vendor directory is deleted. This deletion makes up
most of the commit. The move of files into src makes up the next
significant proportion of changes.

Addition of CHANGELOG.md in theme root. This is checked for a versioned
entry before the package is created.

New release creation system - use with grunt release:{type} where {type}
is major, minor or patch. Not included in this pass is the option to
automatically bump version numbers. Zip files are now named as
fortytwo-x.y.z.zip where x.y.z is the release version number. Inside the
zip is a folder called fortytwo, meaning the zip is compatible with
the WP Upload Theme feature. The zip file is saved to the dist directory
and this is cleared out on each release task invocation.

Current version in package.json set to less than 1.0.0, so that next
release can be release:major.

Directies added to .gitignore so that now bower assets are committed,
and no (future) composer directories, except for the five PHP files
in the wpthumb directory.

Fixes #168