Dynamically require widget files
Moves the data for which widget files are loaded into functions.php
and out of the lib/ directory.

New widgets can be added by dropping the files into subdirectory of
lib/widgets, then adding support for it in functions.php.

Also means that base widget file is only required if any widgets are supported, and wpthumb library is only required if responsive slider widget is supported.