Add JSCS task and config
JSHint is moving away from code standards and more towards highlighting
bad coding practices. It recommends JSCS as the replacement for
detecting whitespace, alignment and other code standards issues.

WP core hasn't yet got on to this yet, so there's no existin config to
go from. However, the JavaScript code standards in the WP handbook do
give quite clear outlines on what is expected. Using that, and going
through each possible rule in JSCS, I've come up with a config that I
think works, and would be good for core (albeit core doesn't follow it's
own standards).

The top section are rules that agree with jQuery standard, which is what
WP generally follows. This was found in the jquery preset for JSCS. We
can't use the jquery preset and then extend it, as it includes
requiresSpacesInsideArrayBrackets: true, which we don't follow without
exceptions, and there's no way I've found to nullify that rule, hence
the need to duplicate some of the rules.

The second section are rules which match up to the existing WP JS code
standards in the handbook.

The last section are rules that aren't in WP JS code standards, but I
think should be. These could always be counted as additionalRules in the
context of the grunt task, should WP core decide not to include them.

See #211