Improve FortyTwo_Walker_Nav_Menu
Some of the code was too different from Walker_Nav_Menu, and
was jumping to using strings too early. This brings start_el() more back
in line with the default method, and re-applies the nav_menu_link_attributes
filter from core.

The start_lvl() method wasn't doing anything different than what was already
in Walker_Nav_Menu. Creating an array of one item, then trying to
implode it into a string, is just the same as giving the plain string.